services/kubernetes/pyproject.toml
[project]
name = "kubernetes"
version = "0.1.0"
description = ""
authors = [{ name = "Tobias Henkel", email = "tobias.henkel@bmw.de" }]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "pulumi==3.153.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.10.3",
    "pip>=24.3.1",
    "pulumi-cloudflare>=5.44.0",
    "pulumi-onepassword>=1.1.3",
    "pulumi-proxmoxve>=6.18.0",
    "pulumiverse-talos>=0.5.1",
    "pulumi-command>=1.0.1",
    "requests>=2.32.3",
    "pulumi-kubernetes>=4.19.0",
]

[dependency-groups]
dev = [
    "distlib>=0.3.9",
    "filelock>=3.16.1",
    "platformdirs>=4.3.6",
    "pre-commit>=4.0.1",
    "pyright>=1.1.390",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.2",
    "yamllint>=1.35.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [".venv"]

[tool.ruff.lint]
extend-select = [
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "ERA",    # flake8-eradicate (commented out code)
    "G",      # flake8-logging-format
    "I",      # isort
    "ISC",    # flake8-implicit-str-concat
    "PL",     # Pylint
    "RET",    # flake8-return
    "RUF100", # unused noqa
    "SLF",    # flake8-self
]
ignore = [
    "ISC001",  # single line string concat
    "PLR0912", # too many branches
    "PLR0913", # too many arguments
    "PLR0915", # too many statements
    "PLR2004", # magic value used in comparison
    "PLR1711", # useless return
]

[tool.ruff.lint.isort]
lines-between-types = 1

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "SLF001", # allow access to private members in tests
]

[tool.ruff.format]
quote-style = "single"

[tool.pyright]
deprecateTypingAliases = true
useLibraryCodeForTypes = true
include = ["src"]
exclude = [
    "__pycache__",
    ".venv",
    ".vscode",
    "build",
    "dist",
    "node_modules",
    "venv",
]
reportUnnecessaryTypeIgnoreComment = true
venv = ".venv"
venvPath = "."
extraPaths = ["../deploy-base/src"]

[tool.pulumi]
config-root-model = "kubernetes.config:PulumiConfigRoot"
